receipt_dynamo/requirements-dev.txt
pylint-pytest 